update examples to follow the new csg region interface
