+++ examples/fixed_source/azurv1_pl_super/input.py
@@ -43,7 +43,7 @@ mcdc.tally.mesh_tally(
 )
 
 # Setting
-mcdc.setting(N_particle=1e5)
+mcdc.setting(N_particle=1e2)
 
 # Run
 mcdc.run()
