+++ examples/fixed_source/cooper2/input.py
@@ -25,10 +25,10 @@ sy2 = mcdc.surface("plane-y", y=2.0)
 sy3 = mcdc.surface("plane-y", y=4.0, bc="vacuum")
 
 # Set cells
-mcdc.cell([+sx1, -sx2, +sy1, -sy2], m_room)
-mcdc.cell([+sx1, -sx4, +sy2, -sy3], m_room)
-mcdc.cell([+sx3, -sx4, +sy1, -sy2], m_room)
-mcdc.cell([+sx2, -sx3, +sy1, -sy2], m_barrier)
+mcdc.cell(+sx1 & -sx2 & +sy1 & -sy2, m_room)
+mcdc.cell(+sx1 & -sx4 & +sy2 & -sy3, m_room)
+mcdc.cell(+sx3 & -sx4 & +sy1 & -sy2, m_room)
+mcdc.cell(+sx2 & -sx3 & +sy1 & -sy2, m_barrier)
 
 # =============================================================================
 # Set source
@@ -42,7 +42,9 @@ mcdc.source(x=[0.0, 1.0], y=[0.0, 1.0], isotropic=True)
 # =============================================================================
 
 # Tally: cell-average and cell-edge angular fluxes and currents
-mcdc.tally(scores=["flux"], x=np.linspace(0.0, 4.0, 40), y=np.linspace(0.0, 4.0, 40))
+mcdc.tally.mesh_tally(
+    scores=["flux"], x=np.linspace(0.0, 4.0, 40), y=np.linspace(0.0, 4.0, 40)
+)
 
 # Setting
 mcdc.setting(N_particle=1e2)
