+++ examples/fixed_source/cooper2/process.py
@@ -5,12 +5,12 @@ import numpy as np
 
 # Load result
 with h5py.File("output.h5", "r") as f:
-    x = f["tally/grid/x"][:]
+    x = f["tallies/mesh_tally_0/grid/x"][:]
     dx = [x[1:] - x[:-1]][-1]
     x_mid = 0.5 * (x[:-1] + x[1:])
 
-    phi = f["tally/flux/mean"][:]
-    phi_sd = f["tally/flux/sdev"][:]
+    phi = f["tallies/mesh_tally_0/flux/mean"][:]
+    phi_sd = f["tallies/mesh_tally_0/flux/sdev"][:]
 
 
 # Plot result
