+++ examples/fixed_source/cooper2_iqmc/process.py
@@ -8,7 +8,7 @@ with h5py.File("output.h5", "r") as f:
     meshy = f["iqmc/grid/y"][:]
     dx = [meshx[1:] - meshx[:-1]][-1]
     x_mid = 0.5 * (meshx[:-1] + meshx[1:])
-    phi = f["iqmc/tally/flux"][:]
+    phi = f["iqmc/tally/flux/mean"][:]
 
     f.close()
 
@@ -17,11 +17,10 @@ with h5py.File("output.h5", "r") as f:
 # =============================================================================
 X, Y = np.meshgrid(x_mid, x_mid)
 Z = np.log10(np.abs(phi / phi.min()))
-fig, ax = plt.subplots(subplot_kw={"projection": "3d"}, dpi=300, figsize=(12, 10))
+fig, ax = plt.subplots(subplot_kw={"projection": "3d"})
 ax.plot_surface(Y, X, Z, edgecolor="b", color="white", linewidth=0.5)
 ax.set_xlabel("x")
 ax.set_ylabel("y")
 ax.set_zlabel(r"log($\phi$)", rotation=180)
 
-ax.view_init(elev=15, azim=20)
 plt.show()
