+++ examples/fixed_source/slab_absorbium/input.py
@@ -19,9 +19,9 @@ s3 = mcdc.surface("plane-z", z=4.0)
 s4 = mcdc.surface("plane-z", z=6.0, bc="vacuum")
 
 # Set cells
-mcdc.cell([+s1, -s2], m2)
-mcdc.cell([+s2, -s3], m3)
-mcdc.cell([+s3, -s4], m1)
+mcdc.cell(+s1 & -s2, m2)
+mcdc.cell(+s2 & -s3, m3)
+mcdc.cell(+s3 & -s4, m1)
 
 # =============================================================================
 # Set source
@@ -35,8 +35,8 @@ mcdc.source(z=[0.0, 6.0], isotropic=True)
 # =============================================================================
 
 # Tally: cell-average fluxes and currents
-mcdc.tally(
-    scores=["flux", "current"],
+mcdc.tally.mesh_tally(
+    scores=["flux"],
     z=np.linspace(0.0, 6.0, 61),
     mu=np.linspace(-1.0, 1.0, 32 + 1),
 )
