+++ examples/fixed_source/slab_reed_dd/input.py
@@ -42,7 +42,7 @@ mcdc.source(z=[5.0, 6.0], isotropic=True, prob=0.5)
 # Set tally, setting, and run mcdc
 # =============================================================================
 
-mcdc.tally.mesh_tally(scores=["flux"], z=np.linspace(0.0, 8.0, 801))
+mcdc.tally.mesh_tally(scores=["flux"], z=np.linspace(0.0, 8.0, 81))
 
 # Setting
 mcdc.setting(N_particle=5000)
