+++ examples/fixed_source/slab_reed_iqmc/input.py
@@ -37,10 +37,10 @@ mcdc.cell(+s7 & -s8, m4)
 # =============================================================================
 # iQMC Parameters
 # =============================================================================
-N = 10000
+N = 1000
 Nx = 64
 maxit = 20
-tol = 1e-4
+tol = 1e-1
 x = np.linspace(-8, 8, num=Nx + 1)
 solver = "gmres"
